+++ src/index.js
@@ -1,5 +1,6 @@
 import { version } from "../package.json";
-import Local from "./LocalStorage.js";
+import Local from "./Local.js";
+import Session from "./Session.js";
 
 /**
  * Get version number
@@ -11,6 +12,7 @@ function getVersion() { return version; }
 const Guardar = {
 	getVersion,
 	Local: new Local(),
+	Session: new Session(),
 };
 
 export default Guardar;