+++ types/index.d.ts
@@ -2,6 +2,7 @@ export default Guardar;
 declare namespace Guardar {
     export { getVersion };
     export const Local: Local;
+    export const Session: Session;
 }
 /**
  * Get version number
@@ -9,4 +10,5 @@ declare namespace Guardar {
  * @returns {string} string of current version
  */
 declare function getVersion(): string;
-import Local_1 from "./LocalStorage.js";
+import Local_1 from "./Local.js";
+import Session_1 from "./Session.js";